dom/base/test/meta_viewport/test_meta_viewport5.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>meta viewport test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="user-scalable=NO">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>user-scalable=NO</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function test1() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      is(info.allowZoom, true, "user-scalable values are case-sensitive; 'NO' is not valid");
    });
  </script>
</body>
</html>
